Highlight grain row when point is picked
This almost works. A row is highlighted when a point is picked,
and this is based upon the `event.ind` that gets passed to the
event callback when a point is picked. In theory, this index
should match up with the index of the data that was used to
generate the point (see [here](https://matplotlib.org/3.3.3/gallery/event_handling/pick_event_demo.html) for an example). However,
in our case, the index does not seem to match the data. I'm
not sure why. It could be a bug on matplotlib's side.

Anyways, if we can fix that, this will be mostly ready to go.
The one other thing we should check is what happens if multiple
points are selected when the user clicks (which one gets chosen?).

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>